get connection functions from utils.py
