+++ sync.py
@@ -21,6 +21,7 @@ from requests.adapters import HTTPAdapter, Retry
 import citygeo_secrets
 from common import *
 from config import *
+from utils import *
 
 #print(os.environ["ORA_TZFILE"])
